chore: move properties to metadata API
