+++ apps/api/routes/internal/property/index.ts
@@ -1,4 +1,4 @@
-import v2PropertyRouter from '../../mgmt/v2/property';
+import v2PropertyRouter from '../../metadata/v2/property';
 
-// All endpoints are the same as the v2 mgmt endpoints
+// All endpoints are the same as the v2 metadata endpoints
 export default v2PropertyRouter;
