+++ apps/api/routes/metadata/v2/index.ts
@@ -3,6 +3,7 @@ import { openapiMiddleware } from '@/middleware/openapi';
 import { Router } from 'express';
 import associationType from './association_type';
 import object from './object';
+import property from './property';
 
 export default function init(app: Router): void {
   const v2Router = Router();
@@ -12,6 +13,7 @@ export default function init(app: Router): void {
 
   associationType(v2Router);
   object(v2Router);
+  property(v2Router);
 
   app.use('/v2', v2Router);
 }
