+++ apps/api/routes/mgmt/v2/index.ts
@@ -1,13 +1,12 @@
 import { apiKeyHeaderMiddleware } from '@/middleware/api_key';
 import { openapiMiddleware } from '@/middleware/openapi';
 import { Router } from 'express';
+import provider from '../../metadata/v2/property';
 import customer from './customer';
 import destination from './destination';
 import entity from './entity';
 import entityMapping from './entity_mapping';
 import fieldMapping from './field_mapping';
-import property from './property';
-import provider from './provider';
 import schema from './schema';
 import sync from './sync';
 import syncConfig from './sync_config';
@@ -29,7 +28,6 @@ export default function init(app: Router): void {
   sync(v2Router);
   syncRun(v2Router);
   fieldMapping(v2Router);
-  property(v2Router);
   entity(v2Router);
   entityMapping(v2Router);
