+++ docs/docs/api/v2/metadata/list-properties.api.mdx
@@ -5,9 +5,9 @@ description: "List properties"
 sidebar_label: "List properties"
 hide_title: true
 hide_table_of_contents: true
-api: {"operationId":"listProperties","tags":["Properties"],"security":[{"x-api-key":[]}],"parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"string","enum":["common","standard","custom"]}},{"name":"name","in":"query","required":true,"schema":{"type":"string","description":"The name of the object for which you want to list properties."}},{"name":"x-customer-id","in":"header","schema":{"type":"string","example":"my-customer-1"},"description":"The customer ID that uniquely identifies the customer in your application","required":true},{"name":"x-provider-name","in":"header","schema":{"type":"string","example":"salesforce"},"description":"The provider name","required":true}],"responses":{"200":{"description":"List of properties","content":{"application/json":{"schema":{"type":"object","properties":{"properties":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","description":"The name of the property as it appears in the API.","example":"first_name"},"label":{"type":"string","description":"The human-readable name of the property.","example":"First Name"}},"required":["id","label"],"title":"property"}}},"required":["properties"]}}}}},"description":"List properties","method":"get","path":"/properties","servers":[{"url":"https://api.supaglue.io/mgmt/v2","description":"Supaglue API"}],"securitySchemes":{"x-api-key":{"type":"apiKey","name":"x-api-key","in":"header","description":"API key to allow developers to access the API"}},"info":{"version":"0.13.4","title":"Management API","contact":{"name":"Supaglue","email":"docs@supaglue.com","url":"https://supaglue.com"},"description":"#### Introduction\n\nWelcome to the Management API documentation. You can use this API to manage customer integrations and connections.\n\n[![Run in Postman](https://run.pstmn.io/button.svg)](https://www.postman.com/supaglue/workspace/supaglue-public/overview)\n\n#### Base API URL\n\n```\nhttps://api.supaglue.io/mgmt/v2\n```\n"},"postman":{"name":"List properties","description":{"type":"text/plain"},"url":{"path":["properties"],"host":["{{baseUrl}}"],"query":[{"disabled":false,"description":{"content":"(Required) ","type":"text/plain"},"key":"type","value":""},{"disabled":false,"description":{"content":"(Required) ","type":"text/plain"},"key":"name","value":""}],"variable":[]},"header":[{"disabled":false,"description":{"content":"(Required) The customer ID that uniquely identifies the customer in your application","type":"text/plain"},"key":"x-customer-id","value":""},{"disabled":false,"description":{"content":"(Required) The provider name","type":"text/plain"},"key":"x-provider-name","value":""},{"key":"Accept","value":"application/json"}],"method":"GET","auth":{"type":"apikey","apikey":[{"type":"any","value":"x-api-key","key":"key"},{"type":"any","value":"<API Key>","key":"value"},{"type":"any","value":"header","key":"in"}]}}}
+api: {"operationId":"listProperties","tags":["Properties"],"security":[{"x-api-key":[]}],"parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"string","enum":["common","standard","custom"]}},{"name":"name","in":"query","required":true,"schema":{"type":"string","description":"The name of the object for which you want to list properties."}},{"name":"x-customer-id","in":"header","schema":{"type":"string","example":"my-customer-1"},"description":"The customer ID that uniquely identifies the customer in your application","required":true},{"name":"x-provider-name","in":"header","schema":{"type":"string","example":"salesforce"},"description":"The provider name","required":true}],"responses":{"200":{"description":"List of properties","content":{"application/json":{"schema":{"type":"object","properties":{"properties":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","description":"The name of the property as it appears in the API.","example":"first_name"},"label":{"type":"string","description":"The human-readable name of the property.","example":"First Name"}},"required":["id","label"],"title":"property"}}},"required":["properties"]}}}}},"description":"List properties","method":"get","path":"/properties","servers":[{"url":"https://api.supaglue.io/metadata/v2","description":"Supaglue API"}],"securitySchemes":{"x-api-key":{"type":"apiKey","name":"x-api-key","in":"header","description":"API key to allow developers to access the API"}},"info":{"version":"0.13.1","title":"Metadata API","contact":{"name":"Supaglue","email":"docs@supaglue.com","url":"https://supaglue.com"},"description":"#### Introduction\n\nWelcome to the Metadata API documentation. You can use this API to operate on Object metadata in third-party providers.\n\n[![Run in Postman](https://run.pstmn.io/button.svg)](https://www.postman.com/supaglue/workspace/supaglue-public/overview)\n\n#### Base API URL\n```\nhttps://api.supaglue.io/metadata/v2\n```\n"},"postman":{"name":"List properties","description":{"type":"text/plain"},"url":{"path":["properties"],"host":["{{baseUrl}}"],"query":[{"disabled":false,"description":{"content":"(Required) ","type":"text/plain"},"key":"type","value":""},{"disabled":false,"description":{"content":"(Required) ","type":"text/plain"},"key":"name","value":""}],"variable":[]},"header":[{"disabled":false,"description":{"content":"(Required) The customer ID that uniquely identifies the customer in your application","type":"text/plain"},"key":"x-customer-id","value":""},{"disabled":false,"description":{"content":"(Required) The provider name","type":"text/plain"},"key":"x-provider-name","value":""},{"key":"Accept","value":"application/json"}],"method":"GET","auth":{"type":"apikey","apikey":[{"type":"any","value":"x-api-key","key":"key"},{"type":"any","value":"<API Key>","key":"value"},{"type":"any","value":"header","key":"in"}]}}}
 sidebar_class_name: "get api-method"
-info_path: api/v2/mgmt/management-api
+info_path: api/v2/metadata/metadata-api
 custom_edit_url: null
 ---
 
@@ -26,7 +26,7 @@ import TabItem from "@theme/TabItem";
 
 <MethodEndpoint method={"get"} path={"/properties"}></MethodEndpoint>
 
-<SecuritySchemes infoPath={"api/v2/mgmt/management-api"}></SecuritySchemes>
+<SecuritySchemes infoPath={"api/v2/metadata/metadata-api"}></SecuritySchemes>
