+++ openapi/v2/metadata/components/schemas/objects/property.yaml
