+++ openapi/v2/metadata/paths/properties.yaml
