--- openapi/v2/mgmt/components/schemas/list_properties.yaml
@@ -1,14 +1,0 @@
-type: object
-properties:
-  type:
-    type: string
-    enum:
-      - common
-      - standard
-      - custom
-  name:
-    type: string
-    example: object_name
-required:
-  - type
-  - name
